simplesdr/requirements.txt
bottle==0.12.5
pyrtlsdr==0.1.1
